Refactor: OpenJTalkドメインの隔離
