+++ voicevox_engine/tts_pipeline/full_context_label.py
@@ -5,6 +5,10 @@ from typing import Self
 
 import pyopenjtalk
 
+from ..model import AccentPhrase as VvAccentPhrase
+from ..model import Mora as VvMora
+from .mora_list import openjtalk_mora2text
+
 
 @dataclass
 class Phoneme:
@@ -454,3 +458,89 @@ def extract_full_context_label(text: str):
     phonemes = [Phoneme.from_label(label=label) for label in labels]
     utterance = Utterance.from_phonemes(phonemes)
     return utterance
+
+
+def mora_to_text(mora: str) -> str:
+    """
+    Parameters
+    ----------
+    mora : str
+        モーラ音素文字列
+    Returns
+    -------
+    mora : str
+        モーラ音素文字列
+    """
+    if mora[-1:] in ["A", "I", "U", "E", "O"]:
+        # 無声化母音を小文字に
+        mora = mora[:-1] + mora[-1].lower()
+    if mora in openjtalk_mora2text:
+        return openjtalk_mora2text[mora]
+    else:
+        return mora
+
+
+def full_context_label_moras_to_moras(full_context_moras: list[Mora]) -> list[VvMora]:
+    """
+    Moraクラスのキャスト (`Mora` -> `model.Mora`)
+    Parameters
+    ----------
+    full_context_moras : List[Mora]
+        モーラ系列
+    Returns
+    -------
+    moras : List[Mora]
+        モーラ系列。音素長・モーラ音高は 0 初期化
+    """
+    return [
+        VvMora(
+            text=mora_to_text("".join([p.phoneme for p in mora.phonemes])),
+            consonant=(mora.consonant.phoneme if mora.consonant is not None else None),
+            consonant_length=0 if mora.consonant is not None else None,
+            vowel=mora.vowel.phoneme,
+            vowel_length=0,
+            pitch=0,
+        )
+        for mora in full_context_moras
+    ]
+
+
+def utterance_to_accent_phrases(utterance: Utterance) -> list[VvAccentPhrase]:
+    """Utteranceインスタンスをアクセント句系列へドメイン変換する"""
+    return [
+        VvAccentPhrase(
+            moras=full_context_label_moras_to_moras(accent_phrase.moras),
+            accent=accent_phrase.accent,
+            pause_mora=(
+                VvMora(
+                    text="、",
+                    consonant=None,
+                    consonant_length=None,
+                    vowel="pau",
+                    vowel_length=0,
+                    pitch=0,
+                )
+                if (
+                    i_accent_phrase == len(breath_group.accent_phrases) - 1
+                    and i_breath_group != len(utterance.breath_groups) - 1
+                )
+                else None
+            ),
+            is_interrogative=accent_phrase.is_interrogative,
+        )
+        for i_breath_group, breath_group in enumerate(utterance.breath_groups)
+        for i_accent_phrase, accent_phrase in enumerate(breath_group.accent_phrases)
+    ]
+
+
+def text_to_accent_phrases(text: str) -> list[VvAccentPhrase]:
+    """日本語テキストからアクセント句系列を生成"""
+    if len(text.strip()) == 0:
+        return []
+
+    # 音素とアクセントの推定
+    utterance = extract_full_context_label(text)
+    if len(utterance.breath_groups) == 0:
+        return []
+
+    return utterance_to_accent_phrases(utterance)
